+++ bstore/tests/parsers_tests.py
@@ -14,10 +14,16 @@ __author__ = 'Kyle M. Douglass'
 __email__ = 'kyle.m.douglass@gmail.com' 
 
 from nose.tools import *
-from bstore  import parsers, database
+
+# Register the test generic
 from bstore  import config
+config.__Registered_Generics__.append('testType')
+
+from bstore  import parsers, database
 from pathlib import Path
 from tifffile import TiffFile
+
+
 
 testDataRoot = Path(config.__Path_To_Test_Data__)
 
@@ -413,6 +419,60 @@ def test_MMParser_TiffFile():
     ok_(isinstance(ds.data, TiffFile))
     assert_equal(ds.data.asarray().shape, (512, 512))
     
+def test_MMParser_ConvertsSpacesToUnderscores():
+    """The MMParser will convert spaces in the prefix to underscores.
+    
+    """
+    acqID       =            1
+    channelID   =       'A647'
+    dateID      =        None,
+    posID       =         (0,) 
+    prefix      = 'my dataset' # Note the space in the name!
+    sliceID     =         None
+    datasetType = 'locResults'    
+    
+    parser = TestParser(prefix, acqID, datasetType,
+                        channelID = channelID, dateID = dateID,
+                        posID = posID, sliceID = sliceID)
+    assert_equal(parser.acqID,                  1)
+    assert_equal(parser.channelID,         'A647')
+    assert_equal(parser.posID,               (0,))
+    assert_equal(parser.prefix,      'my_dataset') # Note the underscore
+    assert_equal(parser.sliceID,             None)
+    assert_equal(parser.datasetType, 'locResults')
+    
+@raises(parsers.GenericTypeError)    
+def test_MMParser_Bad_Generic():
+    """MMParser recognizes when a bad generic type is passed.
+    
+    """
+    inputFile = Path('Cos7_Microtubules_A750_3_MMStack_Pos0_locResults.dat')
+    
+    parser = parsers.MMParser()
+    parser.parseFilename(inputFile,
+                         datasetType ='generic',
+                         genericTypeName = 'bogusType')
+                        
+def test_MMParser_Generic_GetDatabaseAtom():
+    """MMParser returns a generic database atom.
+    
+    """
+    inputFile = Path('Cos7_Microtubules_A750_3_MMStack_Pos0_locResults.dat')
+    
+    parser = parsers.MMParser()
+    parser.parseFilename(inputFile,
+                         datasetType ='generic',
+                         genericTypeName = 'testType')
+    dba = parser.getDatabaseAtom()
+    assert_equal(dba.prefix, 'Cos7_Microtubules')
+    assert_equal(dba.acqID,                    3)
+    assert_equal(dba.datasetType,      'generic')
+    assert_equal(dba.channelID,           'A750')
+    assert_equal(dba.posID,                 (0,))
+    assert_equal(dba.dateID,                None)
+    assert_equal(dba.sliceID,               None)
+    assert_equal(dba.genericTypeName, 'testType')
+    
 def test_FormatMap():
     """FormatMap provides a basic two-way hash table.
     
@@ -452,28 +512,6 @@ def test_FormatMap_Dict_Constructor():
     
     # Tests undefined keys
     assert_equal(testMap['C'], 'C')
-    
-def test_MMParser_ConvertsSpacesToUnderscores():
-    """The MMParser will convert spaces in the prefix to underscores.
-    
-    """
-    acqID       =            1
-    channelID   =       'A647'
-    dateID      =        None,
-    posID       =         (0,) 
-    prefix      = 'my dataset' # Note the space in the name!
-    sliceID     =         None
-    datasetType = 'locResults'    
-    
-    parser = TestParser(prefix, acqID, datasetType,
-                        channelID = channelID, dateID = dateID,
-                        posID = posID, sliceID = sliceID)
-    assert_equal(parser.acqID,                  1)
-    assert_equal(parser.channelID,         'A647')
-    assert_equal(parser.posID,               (0,))
-    assert_equal(parser.prefix,      'my_dataset') # Note the underscore
-    assert_equal(parser.sliceID,             None)
-    assert_equal(parser.datasetType, 'locResults')
     
 def test_SimpleParser_ParseFilename_LocResults():
     """SimpleParser correctly converts files to Datasets/DatabaseAtoms.
@@ -633,7 +671,7 @@ def test_SimpleParser_GetDatabaseAtom_NotInitialized():
     
     """                             
     parser = parsers.SimpleParser()
-    ds = parser.getDatabaseAtom()
+    parser.getDatabaseAtom()
     
 @raises(Exception)
 def test_SimpleParser_BadParse():
@@ -644,3 +682,11 @@ def test_SimpleParser_BadParse():
                              
     parser = parsers.SimpleParser()
     parser.parseFilename(f, datasetType = 'widefieldImage')
+    
+@raises(ValueError)
+def test_Simple_Parser_No_Generics():
+    """Simple Parser will not parse generics.
+    
+    """
+    parser = parsers.SimpleParser()
+    parser.parseFilename('File', datasetType = 'generic')