+++ docs/conf.py
@@ -73,9 +73,9 @@ author = 'Kyle M. Douglass'
 # built documents.
 #
 # The short X.Y version.
-version = '0.2.0'
+version = '0.2.1'
 # The full version, including alpha/beta/rc tags.
-release = '0.2.0'
+release = '0.2.1'
 
 # Napoleon settings
 napoleon_include_private_with_doc = True